pyproject.toml
[tool.poetry]
name = "pypers"
version = "1.0.0"
description = "A utilities module made for personal use, has small snippets which I use in many projects."
authors = ["Divkix <techdroidroot@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://pypers.divkix.me"
repository = "https://github.com/Divkix/pypers"
keywords = ["pypers", "divkix helpers", "python helpers", "py helpers"]
classifiers = [
        "Development Status :: 5 - Production/Stable",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: Implementation :: CPython",
        "Topic :: Software Development :: Libraries",
        "Topic :: Software Development :: Libraries :: Python Modules",
        "Topic :: Utilities",
    ]
include = [
    "LICENSE",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Divkix/pypers/issues"
"Community" = "https://t.me/DivideProjects"

[tool.poetry.dependencies]
python = "^3.7"
motor = "^2.5.1"
dnspython = "^2.2.0"
httpx = "^0.22.0"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pyclean = "^2.0.0"
pre-commit = "^2.17.0"
pylint = "^2.12.2"
mkdocs = "^1.3.0"
mkdocs-material = "^8.2.8"
mkdocs-git-revision-date-plugin = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
